Break out System.Resources classes
Breaks out System.Resources classes into separate code files.
