+++ src/System.Windows.Forms/src/System/Resources/ResXDataNode.cs
@@ -4,15 +4,12 @@
 
 #nullable disable
 
-using System.Collections;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.ComponentModel.Design;
 using System.Diagnostics;
 using System.Drawing;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Reflection;
 using System.Runtime.Serialization;
 using System.Runtime.Serialization.Formatters.Binary;
@@ -734,327 +731,6 @@ namespace System.Resources
         void ISerializable.GetObjectData(SerializationInfo si, StreamingContext context)
         {
             throw new PlatformNotSupportedException();
-        }
-    }
-
-    internal class DataNodeInfo
-    {
-        internal string Name;
-        internal string Comment;
-        internal string TypeName;
-        internal string MimeType;
-        internal string ValueData;
-        internal Point ReaderPosition; //only used to track position in the reader
-
-        internal DataNodeInfo Clone()
-        {
-            return new DataNodeInfo
-            {
-                Name = Name,
-                Comment = Comment,
-                TypeName = TypeName,
-                MimeType = MimeType,
-                ValueData = ValueData,
-                ReaderPosition = new Point(ReaderPosition.X, ReaderPosition.Y)
-            };
-        }
-    }
-
-    // This class implements a partial type resolver for the BinaryFormatter.
-    // This is needed to be able to read binary serialized content from older
-    // NDP types and map them to newer versions.
-    //
-    internal class ResXSerializationBinder : SerializationBinder
-    {
-        private readonly ITypeResolutionService typeResolver;
-        private readonly Func<Type, string> typeNameConverter;
-
-        internal ResXSerializationBinder(ITypeResolutionService typeResolver)
-        {
-            this.typeResolver = typeResolver;
-        }
-
-        internal ResXSerializationBinder(Func<Type, string> typeNameConverter)
-        {
-            this.typeNameConverter = typeNameConverter;
-        }
-
-        public override Type BindToType(string assemblyName, string typeName)
-        {
-            if (typeResolver == null)
-            {
-                return null;
-            }
-
-            typeName = typeName + ", " + assemblyName;
-
-            Type type = typeResolver.GetType(typeName);
-            if (type == null)
-            {
-                string[] typeParts = typeName.Split(',');
-
-                // Break up the assembly name from the rest of the assembly strong name.
-                // we try 1) FQN 2) FQN without a version 3) just the short name
-                if (typeParts != null && typeParts.Length > 2)
-                {
-                    string partialName = typeParts[0].Trim();
-
-                    for (int i = 1; i < typeParts.Length; ++i)
-                    {
-                        string typePart = typeParts[i].Trim();
-                        if (!typePart.StartsWith("Version=") && !typePart.StartsWith("version="))
-                        {
-                            partialName = partialName + ", " + typePart;
-                        }
-                    }
-                    type = typeResolver.GetType(partialName);
-                    if (type == null)
-                    {
-                        type = typeResolver.GetType(typeParts[0].Trim());
-                    }
-                }
-            }
-
-            // Binder couldn't handle it, let the default loader take over.
-            return type;
-        }
-
-        //
-        // Get the multitarget-aware string representation for the give type.
-        public override void BindToName(Type serializedType, out string assemblyName, out string typeName)
-        {
-            // Normally we don't change typeName when changing the target framework,
-            // only assembly version or assembly name might change, thus we are setting
-            // typeName only if it changed with the framework version.
-            // If binder passes in a null, BinaryFormatter will use the original value or
-            // for un-serializable types will redirect to another type.
-            // For example:
-            //
-            // Encoding = Encoding.GetEncoding("shift_jis");
-            // public Encoding Encoding { get; set; }
-            // property type (Encoding) is abstract, but the value is instantiated to a specific class,
-            // and should be serialized as a specific class in order to be able to instantiate the result.
-            //
-            // another example are singleton objects like DBNull.Value which are serialized by System.UnitySerializationHolder
-            typeName = null;
-            if (typeNameConverter != null)
-            {
-                string assemblyQualifiedTypeName = MultitargetUtil.GetAssemblyQualifiedName(serializedType, typeNameConverter);
-                if (!string.IsNullOrEmpty(assemblyQualifiedTypeName))
-                {
-                    int pos = assemblyQualifiedTypeName.IndexOf(',');
-                    if (pos > 0 && pos < assemblyQualifiedTypeName.Length - 1)
-                    {
-                        assemblyName = assemblyQualifiedTypeName.Substring(pos + 1).TrimStart();
-                        string newTypeName = assemblyQualifiedTypeName.Substring(0, pos);
-                        if (!string.Equals(newTypeName, serializedType.FullName, StringComparison.InvariantCulture))
-                        {
-                            typeName = newTypeName;
-                        }
-                        return;
-                    }
-                }
-            }
-
-            base.BindToName(serializedType, out assemblyName, out typeName);
-        }
-    }
-
-    internal class AssemblyNamesTypeResolutionService : ITypeResolutionService
-    {
-        private AssemblyName[] names;
-        private Hashtable cachedAssemblies;
-        private Hashtable cachedTypes;
-
-        private static readonly string s_dotNetPath = Path.Combine(Environment.GetEnvironmentVariable("ProgramFiles"), "dotnet\\shared");
-        private static readonly string s_dotNetPathX86 = Path.Combine(Environment.GetEnvironmentVariable("ProgramFiles(x86)"), "dotnet\\shared");
-
-        internal AssemblyNamesTypeResolutionService(AssemblyName[] names)
-        {
-            this.names = names;
-        }
-
-        public Assembly GetAssembly(AssemblyName name)
-        {
-            return GetAssembly(name, true);
-        }
-
-        //
-        public Assembly GetAssembly(AssemblyName name, bool throwOnError)
-        {
-            Assembly result = null;
-
-            if (cachedAssemblies == null)
-            {
-                cachedAssemblies = Hashtable.Synchronized(new Hashtable());
-            }
-
-            if (cachedAssemblies.Contains(name))
-            {
-                result = cachedAssemblies[name] as Assembly;
-            }
-
-            if (result == null)
-            {
-                result = Assembly.Load(name.FullName);
-                if (result != null)
-                {
-                    cachedAssemblies[name] = result;
-                }
-                else if (names != null)
-                {
-                    foreach (AssemblyName asmName in names.Where(an => an.Equals(name)))
-                    {
-                        try
-                        {
-                            result = Assembly.LoadFrom(GetPathOfAssembly(asmName));
-                            if (result != null)
-                            {
-                                cachedAssemblies[asmName] = result;
-                            }
-                        }
-                        catch
-                        {
-                            if (throwOnError)
-                            {
-                                throw;
-                            }
-                        }
-                    }
-                }
-            }
-
-            return result;
-        }
-
-        public string GetPathOfAssembly(AssemblyName name)
-        {
-            return name.CodeBase;
-        }
-
-        public Type GetType(string name)
-        {
-            return GetType(name, true);
-        }
-
-        public Type GetType(string name, bool throwOnError)
-        {
-            return GetType(name, throwOnError, false);
-        }
-
-        public Type GetType(string name, bool throwOnError, bool ignoreCase)
-        {
-            Type result = null;
-
-            // Check type cache first
-            if (cachedTypes == null)
-            {
-                cachedTypes = Hashtable.Synchronized(new Hashtable(StringComparer.Ordinal));
-            }
-
-            if (cachedTypes.Contains(name))
-            {
-                result = cachedTypes[name] as Type;
-                return result;
-            }
-
-            // Missed in cache, try to resolve the type from the reference assemblies.
-            if (name.IndexOf(',') != -1)
-            {
-                result = Type.GetType(name, false, ignoreCase);
-            }
-
-            if (result == null && names != null)
-            {
-                //
-                // If the type is assembly qualified name, we sort the assembly names
-                // to put assemblies with same name in the front so that they can
-                // be searched first.
-                int pos = name.IndexOf(',');
-                if (pos > 0 && pos < name.Length - 1)
-                {
-                    string fullName = name.Substring(pos + 1).Trim();
-                    AssemblyName assemblyName = null;
-                    try
-                    {
-                        assemblyName = new AssemblyName(fullName);
-                    }
-                    catch
-                    {
-                    }
-
-                    if (assemblyName != null)
-                    {
-                        List<AssemblyName> assemblyList = new List<AssemblyName>(names.Length);
-                        foreach (AssemblyName asmName in names)
-                        {
-                            if (string.Compare(assemblyName.Name, asmName.Name, StringComparison.OrdinalIgnoreCase) == 0)
-                            {
-                                assemblyList.Insert(0, asmName);
-                            }
-                            else
-                            {
-                                assemblyList.Add(asmName);
-                            }
-                        }
-                        names = assemblyList.ToArray();
-                    }
-                }
-
-                // Search each reference assembly
-                foreach (AssemblyName asmName in names)
-                {
-                    Assembly asm = GetAssembly(asmName, false);
-                    if (asm != null)
-                    {
-                        result = asm.GetType(name, false, ignoreCase);
-                        if (result == null)
-                        {
-                            int indexOfComma = name.IndexOf(',');
-                            if (indexOfComma != -1)
-                            {
-                                string shortName = name.Substring(0, indexOfComma);
-                                result = asm.GetType(shortName, false, ignoreCase);
-                            }
-                        }
-                    }
-
-                    if (result != null)
-                    {
-                        break;
-                    }
-                }
-            }
-
-            if (result == null && throwOnError)
-            {
-                throw new ArgumentException(string.Format(SR.InvalidResXNoType, name));
-            }
-
-            if (result != null)
-            {
-                // Only cache types from the shared framework  because they don't need to update.
-                // For simplicity, don't cache custom types
-                if (IsDotNetAssembly(result.Assembly.Location))
-                {
-                    cachedTypes[name] = result;
-                }
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        ///  This is matching %windir%\Microsoft.NET\Framework*, so both 32bit and 64bit framework will be covered.
-        /// </summary>
-        private bool IsDotNetAssembly(string assemblyPath)
-        {
-            return assemblyPath != null && (assemblyPath.StartsWith(s_dotNetPath, StringComparison.OrdinalIgnoreCase) || assemblyPath.StartsWith(s_dotNetPathX86, StringComparison.OrdinalIgnoreCase));
-        }
-
-        public void ReferenceAssembly(AssemblyName name)
-        {
-            throw new NotSupportedException();
         }
     }
 }
